% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{get_observations}
\alias{get_observations}
\title{Query Observations}
\usage{
get_observations(is_animal = NULL, german_name = NULL,
  scientific_name = NULL, eu_guid = NULL, id = NULL, taxagroup = NULL,
  year = NULL, from = NULL, to = NULL, bbox = NULL, limit = 1000)
}
\arguments{
\item{is_animal}{logical; should animals (TRUE) or plants (FALSE) be queried.
NULL (default returns both).}

\item{german_name}{character; filter by german name.}

\item{scientific_name}{characher; filter by scienctific name.}

\item{eu_guid}{numeric; filter by eu_guid (see \code{\link{get_taxa}}).}

\item{id}{numeric; filter by taxon id (see \code{\link{get_taxa}}).}

\item{taxagroup}{characher; filter by taxagroup}

\item{year}{numeric; filter by year of observation}

\item{from}{character; earliest date in format DD.MM.YYYY}

\item{to}{character; latest date in format DD.MM.YYYY}

\item{bbox}{numeric; vector of corner of bounding box to filter (ETRS89/UTM zone 32N)
with xmin, ymin, xmax and ymax values}

\item{limit}{numeric; Number of observations to retrieve (max 1000).}
}
\value{
a data.frame with one oberservation per row:
\describe{
  \item{id}{Observation ID}
  \item{bereich}{Observation group}
  \item{status}{Observation status}
  \item{anzahl}{number of observed taxa}
  \item{lat}{latitude (ETRS89/UTM zone 32N)}
  \item{lon}{longitude (ETRS89/UTM zone 32N)}
  \item{datum}{date}
  \item{titel_deutsch}{German name}
  \item{titel_wissenschaftlich}{Scientific Name}
  \item{bemerkung}{comment}
  \item{phaenogramm}{observation suitable for a phenogramm? (e.g. imagines)}
  \item{lanis_rlp_osiris_art_id}{taxon id (see \code{\link{get_taxa}})}
  \item{is_rasterized}{flag if observation has been assigned to midpoint of raster, see note.}
}
}
\description{
Query Observations
}
\note{
Sensible taxa have been assigned to their raster midpoints.
}
\examples{
mil1 <- get_observations(scientific_name='milvus milvus', year = 2017)
head(mil1)
}
