% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_taxa}
\alias{get_taxa}
\title{Query available taxa}
\usage{
get_taxa(is_animal = NULL, taxagroup = NULL, gid = NULL,
  name_regexp_ci = NULL)
}
\arguments{
\item{is_animal}{logical; should animals (TRUE) or plants (FALSE) be queried.
NULL (default returns both).}

\item{taxagroup}{character; filter by german name of OSIRIS taxagroups (e.g. from \code{\link{get_taxagroups}}).}

\item{gid}{numeric; filter by OSIRIS taxagroup ID (e.g. from \code{\link{get_taxagroups}}).}

\item{name_regexp_ci}{character; a Regular Expression  (POSIX, Case Insensitive) to filter german or scientific name}
}
\value{
a data.frame with id, eu_guid, name_wissenschaftlich (scientific name),
    name_deutsch (german name), artengruppe (taxagroup-name), gid (taxagroup id), is_animal
}
\description{
Query available taxa
}
\note{
A maximum of 1000 entries is returned.
}
\examples{
get_taxa()

}
