% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mapview}
\alias{plot_mapview}
\title{Plot a map of observations}
\usage{
plot_mapview(obs)
}
\arguments{
\item{obs}{data.frame with observations as returned by (see \code{\link{get_observations}})}
}
\value{
a interactive map
}
\description{
Plot a map of observations
}
\examples{
df <- get_observations(scientific_name = 'Gonepteryx rhamni', year = 2017)
plot_mapview(df)
}
