% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_phaenogram}
\alias{plot_phaenogram}
\title{Plot a phenogram}
\usage{
plot_phaenogram(obs, grain = c("monthly", "weekly"))
}
\arguments{
\item{obs}{data.frame with observations as returned by (see \code{\link{get_observations}})}

\item{grain}{character; resolution of the pehnogram ('monthly' or 'weekly')}
}
\value{
a ggplot object
}
\description{
Plot a phenogram
}
\examples{
df <- get_observations(scientific_name = 'Gonepteryx rhamni', year = 2017)
plot_phaenogram(df)
plot_phaenogram(df, 'weekly')
}
