% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_taxagroups}
\alias{get_taxagroups}
\title{Query available taxagroups}
\usage{
get_taxagroups(is_animal = NULL, limit = 1000)
}
\arguments{
\item{is_animal}{logical; should animals (TRUE) or plants (FALSE) be queried.
NULL (default returns both).}

\item{limit}{numeric; Number of observations to retrieve (max 1000).}
}
\value{
a data.frame with id (OSIRIS taxagroup ID), name_deutsch (german name), and is_animal.
}
\description{
Query available taxagroups
}
\note{
A maximum of 1000 entries is returned.
}
\examples{
get_taxagroups()
get_taxagroups(is_animal = FALSE)

}
