% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_taxagroups}
\alias{get_taxagroups}
\title{Query available taxagroups}
\usage{
get_taxagroups(is_animal = NULL)
}
\arguments{
\item{is_animal}{logical; should animals (TRUE) or plants (FALSE) be queried.
NULL (default returns both)}
}
\value{
a data.frame with id, name_deutsch (german name), and is_animal
}
\description{
Query available taxagroups
}
\examples{
get_taxagroups()

}
